move home component to platform-get-resources
