+++ extensions/src/platform-get-resources/src/types/project-types.ts
