+++ lib/platform-bible-react/src/preview/pages/components/advanced.component.tsx
@@ -24,7 +24,6 @@ import { columns, data } from './data-sources/data-table-content';
 import { TabDropdownMenuExample } from './advanced/tab-dropdown-menu-example.component';
 import { SettingSidebarContentSearchExamples } from './advanced/settings-sidebar-content-search.example.component';
 import { PlatformMenubarExample } from './advanced/platform-menubar-example.component';
-import { HomeExample } from './advanced/home.example.component';
 import { ScopeSelectorExample } from './advanced/scope-selector-example.component';
 
 export function Compositions() {
@@ -124,10 +123,6 @@ export function Compositions() {
 
         <VerticalTabsContent value="Filter">
           <FilterExample />
-        </VerticalTabsContent>
-
-        <VerticalTabsContent value="Home">
-          <HomeExample />
         </VerticalTabsContent>
 
         <VerticalTabsContent value="Markdown Renderer">
